refactor: 매거진 업데이트
